@@ -20,7 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.6 <3.0.0'
+  sdk: '>=3.0.6 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -31,13 +31,11 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  splashscreen:  
+  flutter_native_splash: ^2.2.16
   flutter_localizations:
       sdk: flutter
   flutx:
     path: ./flutx
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   get: any
   dio: any
