@@ -5,7 +5,7 @@ import 'package:installation/models/site.dart';
 
 class Order {
   late List<Order> data;
-  late bool state;
+  //late bool state;
   late int id;
   late String work_name;
   late Agent? agent;
@@ -14,9 +14,9 @@ class Order {
   late DateTime? date;
   late String time;
   late String connectionType;
-  late int connectionNumber;
-  late String charger;
-  late String status;
+  String? connectionNumber;
+  String? charger;
+  String? status;
   late String status_key;
   late String new_step;
   List<Activity> activity = [];
@@ -36,14 +36,14 @@ class Order {
     customer =
         json['customer'] != null ? Customer.fromJson(json['customer']) : null;
     site = (json['site'] != null ? Site.fromJson(json['site']) : null)!;
-    date = DateTime.tryParse(json['date']);
+    date = DateTime.parse(json['date']);
     time = json['time'];
     connectionType = json['connection_type'];
     connectionNumber = json['connection_number'];
     charger = json['charger'];
     status = json['status'];
     status_key = json['status_key'];
-    state = json['state'];
+    //state = json['state'];
     activity = (json['activity'] as List)
         .map((activityJson) => Activity.fromJson(activityJson))
         .toList();
@@ -52,7 +52,7 @@ class Order {
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['data'] = this.data.map((v) => v.toJson()).toList();
-    data['state'] = state;
+    //data['state'] = state;
     return data;
   }
 }
