@@ -4,7 +4,7 @@ import 'package:get_storage/get_storage.dart';
 
 class Api {
   static final dio = Dio(BaseOptions(
-      baseUrl: 'https://www.quikrev.com/installation/api',
+      baseUrl: 'http://installation.buzz-storm.com/api/',
       receiveDataWhenStatusError: true,
       headers: {
         'Accept': 'application/json',
